fix: load friend when add
